pRF_expect_analysis/requirements.txt
appnope==0.1.4
asttokens==2.4.1
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.37
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.53.0
future==1.0.0
h5py==3.11.0
html5lib==1.1
idna==3.7
imageio==2.34.2
importlib_metadata==7.2.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==5.2.2
matplotlib-inline==0.1.7
networkx==3.2.1
nibabel==5.2.1
nilearn==0.10.4
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
setuptools==70.1.0
shapely==2.0.4
six==1.16.0
statsmodels==0.14.2
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.43.0
zipp==3.19.2
